ammended javascript dates

+++ riverdischarge/riverdash/consumers.py
@@ -119,8 +119,8 @@ def reading_message(message):
 				r = datetime.strptime(rc,"%m/%d/%Y %H:%M")
 
 				devread_device = 'RDM1111'
-				water_level_one = Decimal(float(sensor_height) - (float(sms_data_list[5]) * (0.01))) #convert cm to m
-				water_level_two = Decimal(float(sensor_height) - (float(sms_data_list[6]) * (0.01))) #convert cm to m
+				water_level_one = Decimal(Decimal(sensor_height) - (Decimal(sms_data_list[5]) * Decimal(0.01))) #convert cm to m
+				water_level_two = Decimal(Decimal(sensor_height) - (Decimal(sms_data_list[6]) * Decimal(0.01))) #convert cm to m
 				device_batt = sms_data_list[7]
 				device = Device.objects.get(device_id=devread_device)
 				device.device_battery = device_batt
